code for predicing
The code modified from MapReduce02.py improves the implementation of reduce_func(). Because the Pool.map need reduce_func() has only one parameter, the implementation of reduce_func() is a little complex. In this code, Pool.starmap, which allows the reduce_func() has two parameters,  is applied to simplify the implementation of reduce_func() .